requirements.txt
aiohttp==3.6.1
aiohttp-cors==0.7.0
aiojobs==0.2.2
bintrees==2.1.0
scikit-learn==0.22.1
matplotlib==3.1.2
numpy==1.17.3
psutil==5.6.5
pytest==5.3.5
pytest-datadir==1.3.1
PyYAML==5.3
scipy==1.3.1
statsmodels==0.11.0
